requirements.txt
Django==5.1.6
psycopg2-binary
python-dotenv
whitenoise
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docutils==0.18.1
dotenv==0.9.9
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.10
imagesize==1.4.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pillow==11.2.1
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
setuptools==79.0.0
simplejson==3.20.1
snowballstemmer==2.2.0
Sphinx==7.1.2
sphinx-rtd-dark-mode==1.3.0
sphinx-rtd-theme==1.3.0rc1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-plantuml==0.30
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
